Refactored, fixed.
