+++ TagCloudGenerator/TagCloudGenerator.Tests/packages.config
@@ -1,5 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="FakeItEasy" version="5.5.0" targetFramework="net48" />
   <package id="FluentAssertions" version="5.9.0" targetFramework="net48" />
   <package id="NUnit" version="3.5.0" targetFramework="net45" />
 </packages>
