+++ TagCloudGenerator/TagCloudGenerator/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="Autofac" version="4.9.4" targetFramework="net48" />
   <package id="CommandLineParser" version="2.6.0" targetFramework="net48" />
 </packages>
